+++ app/Http/Controllers/Api/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\Product;
 use App\Http\Requests\Product\AssignTechRequest;
 use App\Http\Requests\Product\CreateProductRequest;
 use App\Http\Requests\Product\UpdateProductRequest;
@@ -18,15 +19,10 @@ class ProductController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
-    {
-        $products = $this->productService->all();
-
-        return response()->json(['message'=>"Get product list successfully", 'data'=>ProductResource::collection($products)],200);
-    }
-
     public function getByAccount(Request $request)
     {
+        $this->authorize('viewAny', Product::class);
+        
         $products = $this->productService->getByAccount($request->user()->id);
 
         return response()->json(['message'=>"Get education list successfully", 'data'=>ProductResource::collection($products)],200);
@@ -37,6 +33,8 @@ class ProductController extends Controller
      */
     public function store(CreateProductRequest $request)
     {
+        $this->authorize('create', Product::class);
+
         $data = $request->validated();
 
         $newProduct = $this->productService->store($request->user()->id, $data);
@@ -50,6 +48,7 @@ class ProductController extends Controller
     public function show(string $id)
     {
         $product = $this->productService->find($id);
+        $this->authorize('view', $product);
 
         return response()->json(['message'=>"Get product successfully", 'data'=>new ProductResource($product)],200);
     }
@@ -60,6 +59,8 @@ class ProductController extends Controller
     public function update(string $product_id, UpdateProductRequest $request)
     {
         $product = $this->productService->find($product_id);
+        $this->authorize('update', $product);
+
         $data = $request->validated();
 
         $updatedProduct = $this->productService->update($data, $product);
@@ -68,12 +69,18 @@ class ProductController extends Controller
     }
 
     public function assignTechs(string $product_id, AssignTechRequest $request){
+        $product = $this->productService->find($product_id);
+        $this->authorize('update', $product);
+        
         $techs = $this->productService->assignTechs($product_id, $request->validated());
 
         return response()->json(['message'=>"Assign techs to a product successfully", 'data' => TechResource::collection($techs)],200);
     }
 
     public function unAssignTech(string $product_id, string $tech_id){
+        $product = $this->productService->find($product_id);
+        $this->authorize('update', $product);
+
         $this->productService->unAssignTech($product_id, $tech_id);
 
         return response()->json(['message'=>"UnAssign a tech to a product successfully"],200);
@@ -84,6 +91,9 @@ class ProductController extends Controller
      */
     public function destroy(string $product_id)
     {
+        $product = $this->productService->find($product_id);
+        $this->authorize('delete', $product);
+
         $this->productService->destroy($product_id);
 
         return response()->json(['message'=>"Delete a product successfully"],200);
