+++ app/Http/Controllers/Api/TechStackController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Models\Tech;
 use App\Http\Requests\Tech\CreateTechStackRequest;
 use App\Http\Requests\Tech\UpdateTechStackRequest;
 use App\Services\TechStackService;
@@ -18,6 +19,8 @@ class TechStackController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Tech::class);
+
         $techStacks = $this->techStackService->all();
 
         return response()->json(['message'=>"Get tech list successfully", 'data'=>TechResource::collection($techStacks)],200);
@@ -28,6 +31,8 @@ class TechStackController extends Controller
      */
     public function store(CreateTechStackRequest $request)
     {
+        $this->authorize('create', Tech::class);
+
         $data = $request->validated();
 
         $newTechStack = $this->techStackService->store($data);
@@ -41,6 +46,7 @@ class TechStackController extends Controller
     public function show(string $id)
     {
         $techStack = $this->techStackService->find($id);
+        $this->authorize('view', $techStack);
 
         return response()->json(['message'=>"Get tech successfully", 'data'=>new TechResource($techStack)],200);
     }
@@ -51,6 +57,8 @@ class TechStackController extends Controller
     public function update(UpdateTechStackRequest $request, string $id)
     {
         $techStack = $this->techStackService->find($id);
+        $this->authorize('update', $techStack);
+
         $data = $request->validated();
 
         $updatedTechStack = $this->techStackService->update($data, $techStack);
@@ -63,6 +71,9 @@ class TechStackController extends Controller
      */
     public function destroy(string $id)
     {
+        $techStack = $this->techStackService->find($id);
+        $this->authorize('delete', $techStack);
+
         $this->techStackService->destroy($id);
 
         return response()->json(['message'=>"Delete a tech successfully"],200);
